+++ apig-20240327/alibabacloud_apig20240327/client.py
@@ -3623,6 +3623,82 @@ class Client(OpenApiClient):
         headers = {}
         return await self.deploy_mcp_server_with_options_async(mcp_server_id, headers, runtime)
 
+    def describe_regions_with_options(
+        self,
+        request: main_models.DescribeRegionsRequest,
+        headers: Dict[str, str],
+        runtime: RuntimeOptions,
+    ) -> main_models.DescribeRegionsResponse:
+        request.validate()
+        query = {}
+        if not DaraCore.is_null(request.language):
+            query['language'] = request.language
+        req = open_api_util_models.OpenApiRequest(
+            headers = headers,
+            query = Utils.query(query)
+        )
+        params = open_api_util_models.Params(
+            action = 'DescribeRegions',
+            version = '2024-03-27',
+            protocol = 'HTTPS',
+            pathname = f'/v1/regions',
+            method = 'POST',
+            auth_type = 'AK',
+            style = 'ROA',
+            req_body_type = 'json',
+            body_type = 'json'
+        )
+        return DaraCore.from_map(
+            main_models.DescribeRegionsResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def describe_regions_with_options_async(
+        self,
+        request: main_models.DescribeRegionsRequest,
+        headers: Dict[str, str],
+        runtime: RuntimeOptions,
+    ) -> main_models.DescribeRegionsResponse:
+        request.validate()
+        query = {}
+        if not DaraCore.is_null(request.language):
+            query['language'] = request.language
+        req = open_api_util_models.OpenApiRequest(
+            headers = headers,
+            query = Utils.query(query)
+        )
+        params = open_api_util_models.Params(
+            action = 'DescribeRegions',
+            version = '2024-03-27',
+            protocol = 'HTTPS',
+            pathname = f'/v1/regions',
+            method = 'POST',
+            auth_type = 'AK',
+            style = 'ROA',
+            req_body_type = 'json',
+            body_type = 'json'
+        )
+        return DaraCore.from_map(
+            main_models.DescribeRegionsResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def describe_regions(
+        self,
+        request: main_models.DescribeRegionsRequest,
+    ) -> main_models.DescribeRegionsResponse:
+        runtime = RuntimeOptions()
+        headers = {}
+        return self.describe_regions_with_options(request, headers, runtime)
+
+    async def describe_regions_async(
+        self,
+        request: main_models.DescribeRegionsRequest,
+    ) -> main_models.DescribeRegionsResponse:
+        runtime = RuntimeOptions()
+        headers = {}
+        return await self.describe_regions_with_options_async(request, headers, runtime)
+
     def export_http_api_with_options(
         self,
         http_api_id: str,
@@ -5305,6 +5381,8 @@ class Client(OpenApiClient):
             body['targetHttpApiId'] = request.target_http_api_id
         if not DaraCore.is_null(request.version_config):
             body['versionConfig'] = request.version_config
+        if not DaraCore.is_null(request.with_gateway_extension):
+            body['withGatewayExtension'] = request.with_gateway_extension
         req = open_api_util_models.OpenApiRequest(
             headers = headers,
             body = Utils.parse_to_map(body)
@@ -5359,6 +5437,8 @@ class Client(OpenApiClient):
             body['targetHttpApiId'] = request.target_http_api_id
         if not DaraCore.is_null(request.version_config):
             body['versionConfig'] = request.version_config
+        if not DaraCore.is_null(request.with_gateway_extension):
+            body['withGatewayExtension'] = request.with_gateway_extension
         req = open_api_util_models.OpenApiRequest(
             headers = headers,
             body = Utils.parse_to_map(body)