+++ apig-20240327/alibabacloud_apig20240327/models/_describe_regions_request.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class DescribeRegionsRequest(DaraModel):
+    def __init__(
+        self,
+        language: str = None,
+    ):
+        self.language = language
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.language is not None:
+            result['language'] = self.language
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('language') is not None:
+            self.language = m.get('language')
+
+        return self
+